pyproject.toml
[build-system]
requires = ["uv_build>=0.10.6,<0.11.0"]
build-backend = "uv_build"

[project]
name = "marker-analyser"
version = "0.1.0"
requires-python = ">=3.11"
description = "A package for loading and basic analysis of optical tweezer marker data."
authors = [{ name = "Sylvia Whittle", email = "sylviwhittle@gmail.com" }]
license = { file = "LICENSE.md" }
dependencies = [
    "pandas",
    "numpy",
    "matplotlib",
    "lumicks.pylake",
    "pydantic>=2.12.3",
    "ipywidgets",
    "pylint>=4.0.2",
]
readme = "README.md"

[dependency-groups]
dev = [
    "pytest",
    "pylint",
    "mypy",
    "pre-commit",
    "ruff>=0.14.2",
]

[tool.ruff]
exclude = [
  "*.ipynb",
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pycache__",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "docs/conf.py",
  "venv",
]
# per-file-ignores = []
line-length = 120

[tool.numpydoc_validation]
checks = [
  "all", # Perform all check except those listed below
  "ES01", # Missing extended summary
  "EX01", # Missing examples section
  "PR10", # Conflicts with black formatting
  "SA01", # Missing See Also section
]
exclude = [  # don't report on objects that match any of these regex
    "\\.undocumented_method$",
    "\\.__repr__$",
    "^test_",
    "^conftest",
    "^conf$",
]
override_SS05 = [  # override SS05 to allow docstrings starting with these words
    "^Process ",
    "^Assess ",
    "^Access ",
]

# add an entry point when needed
# [project.scripts]
# markeranalyser = "marker_analyser.entry_point:entry_point"
